requirements.txt
# Core data science libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# AutoGluon for model training (as mentioned in mma_ai.md)
autogluon.tabular>=1.0.0

# Database
sqlite3  # Built into Python

# OCR dependencies (for OCR_script.py)
pytesseract>=0.3.10
Pillow>=10.0.0

# Optional: visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
